Deleted examples, renamed examples2 to examples.  See issue #152.
